Completes the coding for Exercise 01
MyNatra Output 1
----------------

We have added complete CRUD operations and support for a resource (Blog). This can serve as an example for what creating a resource should produce.

In this case, the command would be:
```
mynatra resource blog title body
```

What would be different if the command were:
```
mynatra resource comment body
```

Some food for thought.