Adds exercise 01
Car List (client-side code)
---------------------------

We detailed each of the views that we want to support for our CRUD client; Index, New, Show, Edit.

Each has example HTML to give a rough idea of what should be displayed for each view.

The first exercise it to write the client-side code that enables these views to function.